# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Landly API
  description: API для платформы генерации AI-лендингов
  version: 1.0.0
  contact:
    email: support@landly.ru
servers:
  - url: http://localhost:8080/v1
    description: Development server (API)
  - url: https://api.landly.ru/v1
    description: Production server (API)
  - url: http://localhost:8080
    description: Development server (Sites)
  - url: https://landlify.ru
    description: Production server (Sites)
...
paths:
...
  /projects/{id}/generate:
    post:
      tags: [generate]
      summary: Сгенерировать лендинг
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                  description: AI промпт для генерации
                payment_url:
                  type: string
                  format: uri
                  description: URL платёжной формы
      responses:
        '200':
          description: Лендинг сгенерирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{id}/preview:
    get:
      tags: [generate]
      summary: Получить preview лендинга
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: JSON-схема лендинга
          content:
            application/json:
              schema:
                type: object
  /projects/{id}/publish:
    post:
      tags: [generate]
      summary: Опубликовать лендинг
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Лендинг опубликован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        niche:
          type: string
        status:
          type: string
          enum: [draft, generated, published]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PublishResponse:
      type: object
      properties:
        subdomain:
          type: string
        public_url:
          type: string
          format: uri
        published_at:
          type: string
          format: date-time
...